Reorganize corner and wedge code into new burton_extras_t class
